// src/types/network/response/car/car.ts
export interface CarResponse {
  is: number
  token: string
  title: string
  time: string
  location: string
  kilometer: number
  year: number
  color: string
  ad_type: string
  model: string
  fuel: string
  engine: string
  chassis: string
  chassis_front: string
  chassis_back: string
  body: string
  insurance: string
  gearbox: string
  can_exchange: string
  price: number
  phone: string
  status: string
  created_at: string
  updated_at: string
  description: string
  appointment: string | null
  seller_type: string
  my_description: string
}