// src/types/domain/car/car.ts
export interface CarForm {
  token: string
  title: string
  time: string
  location: string
  kilometer: number
  year: number
  color: string
  adType: string
  model: string
  fuel: string
  engine: string
  chassis: string
  chassisFront: string
  chassisBack: string
  body: string
  insurance: string
  gearbox: string
  canExchange: string
  price: number
  phone: string
  status: string
  description: string
  appointment: string | null
  sellerType: string
  myDescription: string
}