// src/types/network/payload/car/car.ts
export interface CarPayload {
  token: string
  title: string
  time: string
  location: string
  kilometer: number
  year: number
  color: string
  ad_type: string
  model: string
  fuel: string
  engine: string
  chassis: string
  chassis_front: string
  chassis_back: string
  body: string
  insurance: string
  gearbox: string
  can_exchange: string
  price: number
  phone: string
  status: string
  description: string
  appointment: string | null
  seller_type: string
  my_description: string
}